@prakashsurya 2017-04-26
additionally, it's worth noting that this change effectively pulls in the changes from: https://github.com/openzfs/openzfs/pull/141 .. thus, we may need to revisit the feedback given on that PR, as it'll most likely apply to this as well.